refactor(path): Moved TildeInPathResolver into configuration module.
